# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the Ticket Place App",
        "contact": {
            "email": "ticket.place@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://apid.c4m.mg/ticket-place-app/public",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get List of users",
                "description": "return list of all users",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "404": {
                        "description": "aucun résultat trouvé"
                    }
                }
            }
        },
        "/api/users/{user_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get One User",
                "description": "return the users",
                "operationId": "e10f96743f349605ddec7dd8f1a09324",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to show",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "404": {
                        "description": "aucun user trouvé"
                    },
                    "500": {
                        "description": "erreur interne serveur"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update one user",
                "description": "update information of one user",
                "operationId": "c64f78766f4915182147dd62bcc1e3f3",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fname": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "lname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0345992047"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "administrator"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "404": {
                        "description": "aucun user trouvé"
                    },
                    "400": {
                        "description": "donné incomplet"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete one user",
                "description": "Delete the user",
                "operationId": "8813fccf55a87ec7b1237ecdee00419f",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "404": {
                        "description": "aucun user trouvé"
                    },
                    "500": {
                        "description": "Erreur interne serveur"
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get profile info",
                "description": "return the current user info",
                "operationId": "4373251f8140344279ac32a1ac817cef",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "aucun résultat trouvé"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/my-notifications": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get notifications",
                "description": "return the user's notifications",
                "operationId": "b179ebcb1cb36e75e1defa6a7c262ce5",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "aucun résultat trouvé"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}